requirements.txt
pydub>=0.24.1
python-telegram-bot>=12.8
SpeechRecognition>=3.8.1
